requirements.txt
python-telegram-bot==20.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
python-dotenv==1.0.0
parameterized==0.9.0

